# Startup speed optimizations