--- a/src/openrct2/core/File.cpp
+++ b/src/openrct2/core/File.cpp
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright (c) 2014-2020 OpenRCT2 developers
+ * Copyright (c) 2014-2021 OpenRCT2 developers
  *
  * For a complete list of all authors, please refer to contributors.md
  * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
@@ -122,6 +122,11 @@ namespace File
     uint64_t GetLastModified(const std::string& path)
     {
         return Platform::GetLastModified(path);
+    }
+
+    uint64_t GetSize(std::string_view path)
+    {
+        return Platform::GetFileSize(path);
     }
 } // namespace File