--- a/src/openrct2/object/Object.h
+++ b/src/openrct2/object/Object.h
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright (c) 2014-2020 OpenRCT2 developers
+ * Copyright (c) 2014-2021 OpenRCT2 developers
  *
  * For a complete list of all authors, please refer to contributors.md
  * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
@@ -223,7 +223,7 @@ public:
     }
 
     bool IsAvailable() const;
-    size_t GetSize() const;
+    uint64_t GetSize() const;
     std::unique_ptr<OpenRCT2::IStream> GetStream() const;
 };