This PR consists of 3 changes to the file IO, resulting in a significant speedup to the startup times:

1. `ObjectAsset::GetSize()` obtained the size of the file by reading it all to memory and counting bytes read, then discarding the contents. This was especially wasteful when used on music tracks and took over 400ms during startup on my setup. To remedy this, I introduced a `File::GetSize()` function that queries the size without reading the file contents. Also made `ObjectAsset::GetSize()` return a 64-bit value, because there is no reason why should the file size be truncated in 32-bit builds.
2. Because of these changes, a "seek to end, tell, seek to begin" idiom can be replaced with a proper size check.
3. After the above two changes, `ftell` showed up as the second hottest function when profiling startup, The culprit turned out to be `FileStream::Read` repeatedly checking for EOF before attempting to read. Since this is an exceptional case that should never/rarely happen, I now let the game try to read from the file and throw an exception if it doesn't read enough - hitting EOF while reading is one of the cases where this will happen. This effectively "optimizes" the function for a success case, while previously it was closer to being "optimized" for a failure case.

I also took the liberty of adding this change to a changelog, since halving the startup time is something easily noticeable for the end user. On my machine, startup times **went down from ~830ms to ~380ms**:

Before: `TitleScreen::Load() took 835ms.`
After: `TitleScreen::Load() took 376ms.`